// frontend/src/api/generated/model/dockerPreview.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Gantry Board API
 * AI Agent Orchestration Kanban Board API
 * OpenAPI spec version: 0.1.0
 */
import type { DockerPreviewContainerId } from './dockerPreviewContainerId';
import type { DockerPreviewErrorMessage } from './dockerPreviewErrorMessage';
import type { DockerPreviewPort } from './dockerPreviewPort';
import type { DockerPreviewPreviewUrl } from './dockerPreviewPreviewUrl';
import type { PreviewStatus } from './previewStatus';

export interface DockerPreview {
  container_id?: DockerPreviewContainerId;
  created_at: string;
  error_message?: DockerPreviewErrorMessage;
  id: string;
  port?: DockerPreviewPort;
  preview_url?: DockerPreviewPreviewUrl;
  status: PreviewStatus;
  updated_at: string;
  worktree_name: string;
}